requirements.txt
certifi==2019.11.28
Django==2.2
djangorestframework==3.11.0
pytz==2019.3
sqlparse==0.3.0
django-rest-swagger==2.1.2
coreapi